backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pydexcom==0.2.3
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
aiofiles==23.2.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
cryptography==41.0.7
sqlite3-fts4==1.0.3
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
loguru==0.7.2
praw==7.7.1
openai==1.3.7
huggingface-hub==0.19.4
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.7.2
dateparser==1.2.0
pytz==2023.3
cachetools==5.3.2
tenacity==8.2.3
